Module Two update.
Module Two update. I completed the tutorial. I'm struggling with the lighting, and with the playerstart falling through the floor.  I was able to use "Start here" feature to bypass that issue to make the recording showing the door functioning. I applied multiple matierals, the walls are plain purple wall, and bricks.  The flooring material indoors is the hex tile, and walnut boards. The roof is cement panels. The outside floor is slate. I change the metal to be more copper colored for the decorative struts and glass window frames instead of a the steel grey color. There is two windows, one door, one sunroof. There is 4 glass wall panes, and one glass door. There is 4 chairs, one table, and one statue.