Module 5 Update Continued implementing customizations
Added the foundation for the letter spheres, whole alphabet. Textures and materials are all setup. Next step is to add the pickup items for the vowel spheres, implement a custom trigger for the target space, adding custom firing of sphere picked up, and then registering correctly spelled words, and some sort of yehaw when successful.